fix(stack): typecheck runtime fixture before build
